+++ juicer/spark/operation.tmpl
@@ -4,13 +4,14 @@
 #
 # Auto-generated Spark code from Lemonade Workflow
 # (c) Speed Labs - Departamento de Ciência da Computação
-# (c) Universidade Federal de Minas Gerais
-# More information about Lemonade
+#     Universidade Federal de Minas Gerais
+# More information about Lemonade to be provided
 #
-from pyspark.sql.functions import *
-from pyspark.sql.window import Window
-from pyspark.sql.types import *
+
 from pyspark.sql import SparkSession
+from pyspark.sql.window import Window
+from pyspark.sql.functions import *
+from pyspark.sql.types import *
 
 {%- for instance in instances %}
 {%- if instance.has_code %}
@@ -27,6 +28,10 @@ def {{instance.parameters.task.operation.slug.replace('-', '_')}}_generate_{{ins
     print '#' * 20
     print '== {} =='.format('{{variable}}')
     {{variable}}.show()
+    # {{variable}}.printSchema()
+    schema = {{variable}}.schema
+    for attr in schema:
+        print attr.nullable, attr.metadata
     {%- endfor %}
 
     return {{instance.get_output_names()}}
